refactor: Add filter_ipv4_address custom template filter
This commit adds a new custom template filter called `filter_ipv4_address` to the `custom_filters.py` file in the `frontend/templatetags` directory. The filter takes in a list of network IPv4 addresses and a list of default gateways. It filters the network IPv4 addresses based on the network prefix of each default gateway and returns a string of filtered addresses.

The purpose of this change is to provide a convenient way to filter network IPv4 addresses based on default gateways in Django templates.