feat: Add boolean_icon custom filter
The `custom_filters.py` file in the `templatetags` directory has been updated to include a new custom filter called `boolean_icon`. This filter returns a checkmark icon if the value is True and a cross icon if the value is False. This addition enhances the visual representation of boolean values in the frontend templates.