refactor: Ignore noisy and irrelevant keys in diff analysis
The `views.py` file in the `frontend` directory has been modified to update the `ignore_keys` list in the `activity` function. This change ignores certain keys in the `diff_report` to improve the accuracy and relevance of the diff analysis. The keys `report_datetime_start`, `report_datetime_end`, `slow_test[]`, `uptime_in_seconds`, `uptime_in_days`, and `deleted_file[]` are now excluded from the analysis as they are either always different or not relevant for the user. This refactor enhances the quality of the diff analysis and provides more meaningful insights to the user.